Cleanup

+++ service/src/overlay.rs
@@ -10,25 +10,24 @@ use windows::Win32::{
     UI::WindowsAndMessaging::GetForegroundWindow,
 };
 
-use crate::{config, system_provider, watcher};
+use winapi::um::wingdi::RGB;
+use winapi::um::winuser::{SetLayeredWindowAttributes, LWA_COLORKEY};
 
-//use winapi::
-// The overlay code is adapted from https://github.com/jcdavis/hroverlay, which is released under the
-// Apache 2.0 license (https://raw.githubusercontent.com/jcdavis/hroverlay/main/LICENSE).
+use crate::{config, system_provider, watcher};
 
 #[derive(Default, NwgUi)]
 pub struct Overlay {
-    #[nwg_control(size: (400, 100), position: (400, 0), flags: "POPUP", ex_flags: winapi::um::winuser::WS_EX_TOPMOST|winapi::um::winuser::WS_EX_LAYERED)]
+    #[nwg_control(size: (400, 120), position: (400, 0), flags: "POPUP", ex_flags: winapi::um::winuser::WS_EX_TOPMOST|winapi::um::winuser::WS_EX_LAYERED)]
     #[nwg_events( OnInit: [Overlay::on_init], OnWindowClose: [Overlay::close] )]
     window: nwg::Window,
 
     #[nwg_layout(parent: window, margin: [0,0,0,0], spacing: 0)]
     layout: nwg::GridLayout,
 
-    #[nwg_resource(family: "Ubuntu", size: 100, weight: 700)]
+    #[nwg_resource(family: "Ubuntu", size: 120, weight: 700)]
     font: nwg::Font,
 
-    #[nwg_control(text: "", size: (400, 100), font: Some(&data.font), h_align: HTextAlign::Right, background_color: Some([255, 0, 0]))]
+    #[nwg_control(text: "", size: (400, 120), font: Some(&data.font), h_align: HTextAlign::Right, background_color: Some([255, 0, 0]))]
     #[nwg_layout_item(layout: layout, row: 0, col: 0)]
     time_label: nwg::Label,
 
@@ -41,9 +40,6 @@ pub struct Overlay {
 
 impl Overlay {
     fn on_init(&self) {
-        use winapi::um::wingdi::RGB;
-        use winapi::um::winuser::{SetLayeredWindowAttributes, LWA_COLORKEY};
-
         let notice = self.notice.sender();
         let (sender, receiver) = mpsc::channel();
 
@@ -66,7 +62,7 @@ impl Overlay {
 
                 for p in cfg.overlay.show_pattern.iter() {
                     let delay: i32 = *p;
-                    let sleep = delay.abs();
+                    let sleep = delay.abs() as u64;
 
                     if delay > 0 {
                         *display_text.lock().unwrap() = rcv.clone();
@@ -76,7 +72,7 @@ impl Overlay {
 
                     notice.notice();
 
-                    thread::sleep(Duration::from_secs(sleep as u64));
+                    thread::sleep(Duration::from_secs(sleep));
                 }
 
                 *display_text.lock().unwrap() = "".to_string();
@@ -137,7 +133,6 @@ fn get_position(label_sz: (u32, u32), x_offset: u32, y_offset: u32) -> (i32, i32
         let _res = GetMonitorInfoA(hmnt, &mut minf as _);
     }
 
-    // get boundaries for top left corner of label
     let lft_boundary = minf.rcMonitor.left;
     let rt_boundary = minf.rcMonitor.right - (label_sz.0 as i32);
     let pos_x = ((rt_boundary - lft_boundary) as f32 * x_offset as f32 / 100_f32).floor() as i32;
